% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dotRcpp}
\alias{dotRcpp}
\title{Dot product in Rcpp.}
\usage{
dotRcpp(x1, x2)
}
\arguments{
\item{x1}{numeric vector}

\item{x2}{numeric vector}
}
\value{
dot product
}
\description{
Dot product in Rcpp.
}
