% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{make_list_species_fromR_tocpp}
\alias{make_list_species_fromR_tocpp}
\title{Transform c++ output into R objects (lists and arrays)}
\usage{
make_list_species_fromR_tocpp(list_species_after_landchange)
}
\description{
Processing output from c++
}
