% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{map_environment_1}
\alias{map_environment_1}
\title{Gridded map (temperature)}
\format{
It is basically a matrix that could come from any spreedsheet in .csv or any other format
}
\description{
Each cell indicates the value for the enviromental variable (only integers). Inhabitable cells take -9
}
