% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{map_k_1}
\alias{map_k_1}
\title{Gridded map (K)}
\format{
It is basically a matrix that could come from any spreedsheet in .csv or any other format
}
\description{
Each cell indicates the total number of individuals that can coexist at a give cell. Inhabitable cells take -9
}
