% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_functions.R
\name{compute_PopulationsSpecificFST}
\alias{compute_PopulationsSpecificFST}
\title{Population genetic metrics}
\usage{
compute_PopulationsSpecificFST(
  this_species_pop_info,
  pairwise = FALSE,
  locus = "both"
)
}
\arguments{
\item{this_species_pop_info}{Table that includes, for a number of species, the xy coordinates of each population as well as their allelic frequencies.}

\item{pairwise}{Boolean, should the analysis be done in a pairwise fashion?}

\item{locus}{Select the locus to analysis "selection", "neutral", or "both"}
}
\value{
A list to retrieve Weir and Cockrham Fst and Hudson's Fst.
}
\description{
Computes population genetic metrics using output from run_simulation()
}
\examples{
library(geneclade)
library(DDD)
library(hierfstat)
# load a table which is one of the three elements retrived by run_simulation().
tablespecies_overtime <- get("tablespecies_overtime")
this_time_slice <- tablespecies_overtime$`time: 9.5`
# Take species with ID = 1
this_species <- this_time_slice[which(this_time_slice$Sp_ID == 1),]
# remove columns with uneccesary info (xy coordinates for instance)
this_species_pop_info <- this_species[,-c(1,2,3,4)]
computed_genenetic_metrics <- compute_PopulationsSpecificFST(this_species_pop_info,
                                                            pairwise = FALSE,
                                                            locus="both")
# Hudson's Fst
mean(as.numeric(computed_genenetic_metrics$computed_betas$betaiovl),na.rm=T)
# Weir and Cockrham Fst
computed_genenetic_metrics$wc_fst$FST
}
