% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_R.R
\name{make_advanced_initialization}
\alias{make_advanced_initialization}
\title{Initialisation for a simulation when defaults are not adequate.}
\usage{
make_advanced_initialization(
  number_spp,
  all_x,
  all_y,
  all_traits,
  all_births,
  all_IDs,
  all_parents,
  all_deaths,
  all_ranges,
  all_alleles,
  all_alleles_neutral,
  all_popsize,
  starting_time
)
}
\arguments{
\item{number_spp}{number of species to start the simulation with}

\item{all_x}{Map coordinates (vector) in x for the first populations in the simulation.}

\item{all_y}{Map coordinates (vector) in y for the first populations in the simulation.}

\item{all_traits}{initial trait state, vector, same length than number_spp}

\item{all_births}{birthdate of each species, vector, same length than number_spp}

\item{all_IDs}{ids of each species in the system, vector, same length than number_spp}

\item{all_parents}{vector of the ids for the parents to the species, same length than number_spp}

\item{all_deaths}{vector of record if deaths of species, 0 means alive, same length than number_spp}

\item{all_ranges}{vector of number of cells where each species is present. Currently, the model only handle species starting with range size of one.}

\item{all_alleles}{vector of allelic frequencies for allele a1, b1, c1...to y1.}

\item{all_alleles_neutral}{vector of allelic frequencies for allele a2, b2, c2...to y2.}

\item{all_popsize}{vector of population size for each species, it should be consistent with all_alleles and all_alleles_neutral.}

\item{starting_time}{starting simulated time, it should be consistent with the times in all_births and all_deaths arguments.}
}
\value{
List of variables needed to start a simulation.
}
\description{
Prepares an object to be used as starting point for a simulation. It requieres the user to provide spatial and geographic information of each species.
}
\examples{

number_spp <- 190
all_x <- c(rep(2:20,10))#3#3#10 #  initial population column
all_y <- c(rep(3,19), rep(5,19),rep(7,19),rep(9,19),rep(11,19),
          rep(13,19),rep(15,19),rep(17,19),rep(19,19),rep(21,19))#3 #3#10 # initial population row
all_traits <- c(rep(100,190)) # initial trait value
all_births <- c(rep(0,190))
all_IDs <- c(1:190)
all_parents <- c(rep(0,190))
all_deaths <- c(rep(0,190))
all_ranges <- c(rep(1,190))
all_alleles <- rep(6,15) # allelic frequencies for allele a, b, c, d, and e.
all_alleles_neutral <- rep(6,15)
all_popsize <- c(rep(90,190))#population size for the first population. It should be the same than sum(all_alleles)
starting_time <- 20
make_advanced_initialization (number_spp,
                                        all_x,
                                        all_y,
                                        all_traits,
                                        all_births,
                                        all_IDs,
                                        all_parents,
                                        all_deaths,
                                        all_ranges,
                                        all_alleles,
                                        all_alleles_neutral,
                                        all_popsize,
                                        starting_time)


}
