% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_R.R
\name{run_geneclade_simulation}
\alias{run_geneclade_simulation}
\title{Run geneclade model simulation on a map.}
\usage{
run_geneclade_simulation(
  position_start_x,
  position_start_y,
  advanced_initialization = NULL,
  max_spp,
  simulated_time,
  condition_to_stop = "richness",
  stop_time_after_change = 0,
  time_slices,
  maximum_cycles = 2e+09,
  rate_speciation,
  rate_colonisation,
  rate_extirpation,
  rate_geneflow,
  rate_demographicchange,
  rate_traitevolution,
  rate_mutation,
  alleles_optimum_enviroment,
  percentage_geneflow = 10,
  vicariant_speciation,
  manual_speciation_events_timing = 0,
  growth_only = TRUE,
  unlink_range_to = NULL,
  x_max,
  y_max,
  map_k_1,
  map_environment_1,
  frequency_print_simulationstatus = 1000
)
}
\arguments{
\item{position_start_x}{Map coordinate in x for the first population in the simulation.}

\item{position_start_y}{Map coordinate in y for the first population in the simulation.}

\item{advanced_initialization}{When initilization needs to be different from default one (one sp with one population of size 150, all alles in the same frequency), user needs to input a list created with make_advanced_initialization() function. Default for this argument is NULL}

\item{max_spp}{Maximum expected number of species. Notice that this is to be used along with condition_to_stop = "richness", see below.}

\item{simulated_time}{The time the simulation will run for. The scale of the simulated time is highly dependent on selected rates.}

\item{condition_to_stop}{The simulation will stop when the number of species is met ("richness") or when the simulated time is up ("time"). Default is "richness".}

\item{stop_time_after_change}{Indicates the percentage of time that the simulation will be let run further, considering the time it needed to reach the first equilibirum. Default is 0 which means that the simulation will run until reaching a second equilibrium after map change (see below).}

\item{time_slices}{A vector with the simulated time points at which the current state of the simulation will be saved in memory to retrive at the end of the simulation. Notice that you might need to know more or less the temporal scale of the total simulation.}

\item{maximum_cycles}{Useful parameter to allow a speficic number of events (every cycle has an event e.g., colonization, geneflow) before stopping the simulation.}

\item{rate_speciation}{Per-population rate of speciation rate (lambda)}

\item{rate_colonisation}{Per-population rate of colonisation rate (gamma). Dispersal of individuals of one cell to an adjacent which is below carrying capacity K.}

\item{rate_extirpation}{Per-population rate local extinction rate (mu). For a given species, a population is extirpated from a cell.}

\item{rate_geneflow}{Per-population rate of geneflow (sigma). One population exchange individuals with an adjacent one.}

\item{rate_demographicchange}{Vector of length 2. First element is the per-population rate of change in population size (delta). Population can grow or shrink. Second element is the standard deviation for the normal distribution from which the next population size will be drawn. Example rate_demographicchange <- c(0.01,60)}

\item{rate_traitevolution}{Vector of length 2. First element is the per-population rate of change in trait value. Second element is the standard deviation for the normal distribution from which the next trait value  will be drawn. Example rate_traitevolution <- c(0.01,1). Notice that all individuals of a species have the same trait state.}

\item{rate_mutation}{Per-population rate of mutation (q) where one individual changes its two loci to different ones.}

\item{alleles_optimum_enviroment}{Vector with 25 integer values representing the environmental optimum for each allele from a1 to y1. Notice that if all 25 are the same, the model will feature no differences in local adaptation and will therefore be neutral. These values should make sense with the cell values in the map_environment_1, see below.}

\item{percentage_geneflow}{This arguments regulates how large, in each geneflow event, is the exchange of individuals between two populations. Because it is a percentage, the exchange can be assymetrical. Default is 10.}

\item{vicariant_speciation}{Boolean. Whether or not species speciate by vicariance. False will result in point-mutation speciation.}

\item{manual_speciation_events_timing}{In case it is needed, for some reason, to force the simulation to have speciation events at given times, this vector should provide the time for those events. This means that the length of this vector will indicate the total number of speciation events. This will automatically turn rate_speciation = 0. This option is disable by default (i.e., manual_speciation_events_timing)}

\item{growth_only}{Boolean. Should population fluctiations only lead to population growth? Useful to ensure local saturation of species.}

\item{unlink_range_to}{If one wishes to make speciation, colonization, or both independent from the number of populations use unlink_range_to <- c("colonization","speciation"). We suggest not using this option. Default is NULL.}

\item{x_max}{Length of x axis in the map.}

\item{y_max}{Length of y axis in the map.}

\item{map_k_1}{dataframe to be the map to run simulations on (dimensions x_max and y_max). Inhabitable cells marked with -9. Cell values represent local carrying capacity K.}

\item{map_environment_1}{dataframe to be the map to run simulations on (dimensions x_max and y_max). Inhabitable cells marked with -9. Cell values represent local environmental conditions e.g., temperature. Only integers.}

\item{frequency_print_simulationstatus}{How many cycles the status of the simulation should be printed in console? Notice that the information printed in one cycle might not be very different from the previous one so increasing frequency_print_simulationstatus might be an option. This is highly dependent on selected rates and does not affect simulation behaviour, it is just a visual aid on what is going on. Default is 1000.}
}
\value{
List of three objects: First is a table with all species in the simulation. For each species, the geographic location of each population as well as the number of individuals carrying each allele is indication. The second object is a list a phylogenetic tree for the species in the simulation. The third one is a table with richness and abundance per cell that can be used to make a map. If multiple time slices are requested, there will be multiple species table, phylogenetic trees and richness maps.
}
\description{
Runs a spatially explicit simulation where alleles, populations and species are modelled. Demographic, dispersal and evolutionary processes are simulated in continuous time.
}
\examples{
library(geneclade)
position_start_x <- 17
position_start_y <- 17
max_spp <- 20
simulated_time <- 250
time_slices <- c(10,20)
rate_speciation <- 0.00005
rate_colonisation <- 5
rate_extirpation <- 0.0000001
rate_geneflow <- 0.01
rate_demographicchange <- 1
rate_traitevolution <- 0
rate_mutation <- rate_geneflow/10
alleles_optimum_enviroment <-  rep(9,25)
vicariant_speciation <- TRUE
x_max <- 42
y_max <- 42
map_k_1 <- get("map_k_1")
# one normally loads the map with:
#map_k_1 <- read.table(paste0("k_map_uniform.txt"))
map_environment_1 <- get("map_environment_1")
#map_environment_1 <- read.table(paste0("temperature_map_uniform.txt"))
maximum_cycles <- 200000
frequency_print_simulationstatus <- 1000
output_simulation <- run_geneclade_simulation (position_start_x,
                                    position_start_y,
                                    advanced_initialization = NULL,
                                    max_spp,
                                    simulated_time,
                                    condition_to_stop = "richness",
                                    stop_time_after_change = 0,
                                    time_slices,
                                    maximum_cycles = maximum_cycles,
                                    rate_speciation,
                                    rate_colonisation,
                                    rate_extirpation,
                                    rate_geneflow,
                                    rate_demographicchange,
                                    rate_traitevolution,
                                    rate_mutation,
                                    alleles_optimum_enviroment,
                                    percentage_geneflow = 10,
                                    vicariant_speciation,
                                   manual_speciation_events_timing = 0,
                                    growth_only = TRUE,
                                    unlink_range_to = NULL,
                                    x_max,
                                    y_max,
                                    map_k_1,
                                    map_environment_1,
                                    frequency_print_simulationstatus)
richnessabundancemap_overtime <- output_simulation$output_richnessabundancemap_overtime
phylotrees_overtime <- output_simulation$output_phylotree_overtime
tablespecies_overtime <- output_simulation$output_table_overtime
# For plotting richness
#
# library(ggplot2)
# library(ggpubr)
#
# richness_abundance_map_present <- richnessabundancemap_overtime[[length(richnessabundancemap_overtime)]]
# p1 <- ggplot(richness_abundance_map_present, aes(X, Y)) +                           # Create heatmap with ggplot2
#   # scale_fill_viridis_c(option = "B", direction = 1) +
#   scale_fill_distiller(palette = "Spectral", direction = -1,
#                        name="Total abundance") +
#   geom_tile(aes(fill = (abundance) )) +
#   theme_classic()+
#   theme_void()
#
# p2 <- ggplot(richness_abundance_map_present, aes(X, Y)) +                           # Create heatmap with ggplot2
#   #scale_fill_viridis_c(option = "B", direction = -1) +
#   scale_fill_distiller(palette = "Spectral", direction = -1,
#                        name="Species Richness") +
#   geom_tile(aes(fill = richness )) +
#   theme_classic()+
#   theme_void()
# ggarrange(p1,p2)
}
