% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_R.R
\name{run_simulation}
\alias{run_simulation}
\title{Run geneclade model using rates for processes taking place at ecological and evolutionary time scales in a spatial context.}
\usage{
run_simulation(
  position_start_x,
  position_start_y,
  advanced_initialization = NULL,
  max_spp,
  simulated_time,
  condition_to_stop = "richness",
  maximum_cycles = 2e+09,
  rate_speciation,
  rate_colonisation,
  rate_extirpation,
  rate_geneflow,
  rate_demographicchange,
  rate_traitevolution,
  rate_mutation,
  percentage_geneflow = 10,
  vicariant_speciation,
  x_max,
  y_max,
  map_k_1,
  map_k_2
)
}
\arguments{
\item{position_start_x}{Map coordinate in x for the first population in the simulation.}

\item{position_start_y}{Map coordinate in y for the first population in the simulation.}

\item{advanced_initialization}{When initilization needs to be different from default one (one sp with one population of size 150, all alles in the same frequency), user needs to input a list created with make_advanced_initialization() function. Default for this argument is NULL}

\item{max_spp}{Total expected number of species.}

\item{simulated_time}{The time the simulation will run for. The scale of the simulated time is highly dependent on selected rates.}

\item{condition_to_stop}{The simulation will stop when the number of species is met ("richness") or when the simulated time is up ("time"). Default is "richness".}

\item{maximum_cycles}{Useful parameter to allow a speficic number of events (every cycle has an event e.g., colonization, geneflow) before stopping the simulation.}

\item{rate_speciation}{Per-population rate of speciation rate (lambda)}

\item{rate_colonisation}{Per-population rate of colonisation rate (gamma). Dispersal of individuals of one cell to an adjacent which is below carrying capacity K.}

\item{rate_extirpation}{Per-population rate local extinction rate (mu). For a given species, a population is extirpated from a cell.}

\item{rate_geneflow}{Per-population rate of geneflow (sigma). One population exchange individuals with an adjacent one.}

\item{rate_demographicchange}{Per-population rate of change in population size (delta). Population can grow or shrink.}

\item{rate_traitevolution}{Per-population rate of change in trait value. Their is one value per species.}

\item{rate_mutation}{Per-population rate of mutation (q) where one individual changes its two loci to different ones.}

\item{percentage_geneflow}{This arguments regulates how large, in each geneflow event, is the exchange of individuals between two populations. Because it is a percentage, the exchange can be assymetrical. Default is 10.}

\item{vicariant_speciation}{Boolean. Whether or not species speciate by vicariance. False will result in point-mutation speciation.}

\item{x_max}{Length of x axis in the map}

\item{y_max}{Length of y axis in the map}

\item{map_k_1}{dataframe to be the first map to run simulations on (dimensions x_max and y_max). Inhabitable cells marked with -9. Cell values represent local carrying capacity K.}

\item{map_k_2}{dataframe to be the second map to run simulations on (dimensions x_max and y_max). Inhabitable cells marked with -9. Cell values represent local carrying capacity K. This map will replace map_k_2 at THIS TIME!!!!}
}
\value{
List of species with detailed information on the variables tracked over time, ready for processing with THESE FUNCTIONS.
}
\description{
Runs a spatially explicit simulation where alleles, populations and species are modelled. Demographic, dispersal and evolutionary processes are simulated in continuous time.
}
\examples{
# Example of how to set the arguments for a Maximum Likelihood search.
library(geneclade)
You can use this output in the plotting function: plot_biogeo_reconst()
}
